requirements.txt
rq==1.10.1
redis==4.1.3
pymongo==4.0.1
python-dateutil==2.8.2
boto3==1.20.51
requests==2.27.1
pandas==1.3.3
unidecode==1.3.2
cchardet==2.1.7
kafka-python==2.0.2